Create Accelerometer Plotting Code
This code was used to visualize the data coming from the MPU6050 and the ADS1115.  The MPU6050 gave us data about the motion of the users hand, and the ADS1115 gave us analog voltage values that corresponded to amount of bending from a finger.   By plotting the values from both on a single plot I was able to visualize the data as well as the timing difference and used this information to develop my motion/gesture detection algorithm.